memcg-v1-no-need-for-memcg-locking-for-mglru-fix
remove !rcu_read_lock_held() assertion

  WARNING: CPU: 0 PID: 85 at mm/vmscan.c:3140 folio_update_gen+0x23d/0x250 mm/vmscan.c:3140
  ...

This assertion is incorrect after this patch series that has removed the
charge migration and has removed mem_cgroup_trylock_pages() /
mem_cgroup_unlock_pages() from the caller of this function
(folio_update_gen()).

The assertion was caused by the patch in this thread.  It used to assert
that a folio must be protected from charge migration.  Charge migration is
removed by this series, and as part of the effort, this patch removes the
RCU lock.

Link: https://lkml.kernel.org/r/ZykEtcHrQRq-KrBC@google.com
Link: https://syzkaller.appspot.com/bug?extid=24f45b8beab9788e467e
Link: https://lore.kernel.org/lkml/67294349.050a0220.701a.0010.GAE@google.com/
Signed-off-by: Yu Zhao <yuzhao@google.com>
Cc: Hugh Dickins <hughd@google.com>
Cc: Johannes Weiner <hannes@cmpxchg.org>
Cc: Michal Hocko <mhocko@kernel.org>
Cc: Muchun Song <muchun.song@linux.dev>
Cc: Roman Gushchin <roman.gushchin@linux.dev>
Cc: Shakeel Butt <shakeel.butt@linux.dev>
Cc: Yosry Ahmed <yosryahmed@google.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>